Support negative numbers for popHead and popTail
For popHead, for a negative number of lines, -N , will pop up to the
last N lines. For example, a file with 5 lines called with popHead -2
will pop the first 3 lines.

For popHead, for a negative number of lines, -N, will pop from the tail
up to the Nth line. For example, a file with 5 lines called with popTail
-n2 will pop the last 3 lines.